+++ CHANGELOG.md
@@ -9,3 +9,17 @@ All notable changes to this project will be documented in this file.
 - Designed a TaskManager component
 - Designed a PasswordManager component
 - Designed a WorkoutTracker component
+
+
+
+## 2026.04.24
+
+### Added
+
+- Designed a proof of concept for TaskManager component
+
+### Updated
+
+- Changed design to focus on TaskManager as the main component moving forward
+- Added a Java proof-of-concept with task adding, removing, checking, counting, clearing, and a main method demo
+